Multithreading is now working for tracers

+++ NBodySim/compile_run.sh
@@ -10,9 +10,11 @@ time ./simulator
 
 ffmpeg -f image2 -r 24 -pattern_type glob -i './outputImages/*.png' output.mp4
 
-if [ $(uname) = "Darwin" ]; then
-	open ./output.mp4
-else
-	killall vlc
-	vlc -q ./output.mp4
+if [ $autoplay = true ]; then
+	if [ $(uname) = "Darwin" ]; then
+		open ./output.mp4
+	else
+		killall vlc
+		vlc -q ./output.mp4
+	fi
 fi
